Moved CloudFunction to index file
